Move the discard logger to a new file
We can ensure that building on other os versions will prevent the
manifold from correctly linking to the discard logger.
