+++ worker/syslogger/syslogger_unix.go
@@ -8,7 +8,6 @@ package syslogger
 
 import (
 	"io"
-	"io/ioutil"
 	"log/syslog"
 )
 
@@ -17,22 +16,4 @@ import (
 // can ensure that the underlying logger can be closed when done.
 func NewSyslog(priority Priority, tag string) (io.WriteCloser, error) {
 	return syslog.New(syslog.Priority(priority), tag)
-}
-
-// NewDiscard creates a new WriteCloser that discards all writes and the close
-// is a noop.
-func NewDiscard(priority Priority, tag string) (io.WriteCloser, error) {
-	return nopCloser{
-		Writer: ioutil.Discard,
-	}, nil
-}
-
-// nopCloser is a closer that discards the close request. We can't use the
-// io.NopCloser as that expects a io.Reader.
-type nopCloser struct {
-	io.Writer
-}
-
-func (nopCloser) Close() error {
-	return nil
 }
